refactored changes according to issue
