+++ src/main/java/Application/DTOs/AuctionDTO.java
@@ -1,4 +1,4 @@
-package Application;
+package Application.DTOs;
 
 import java.util.Date;
