+++ src/main/java/Application/DTOs/ItemDTO.java
@@ -1,4 +1,4 @@
-package Application;
+package Application.DTOs;
 
 import java.util.Set;
 import java.util.stream.Collectors;
