+++ src/main/java/Application/ShoppingService.java
@@ -1,6 +1,11 @@
 package Application;
 import java.util.Set;
 import java.util.stream.Collectors;
+
+import Application.DTOs.OrderDTO;
+import Application.utils.Error;
+import Application.utils.Response;
+
 import java.util.Date;
 
 import Domain.Shopping.IShoppingBasketRepository;
