+++ src/main/java/Domain/Shopping/IReceiptRepository.java
@@ -1,7 +1,6 @@
 package Domain.Shopping;
 
 import Domain.Store.Product;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
