+++ src/main/java/Domain/management/Permission.java
@@ -1,4 +1,4 @@
-package Domain;
+package Domain.management;
 
 import java.util.Collections;
 import java.util.HashSet;
