+++ src/test/java/UserServiceTests.java
@@ -4,9 +4,9 @@ import org.junit.Before;
 import org.junit.Test;
 
 import Application.UserService;
+import Application.DTOs.UserDTO;
+import Application.utils.Response;
 import Domain.User.LoginManager;
-import Application.Response;
-import Application.UserDTO;
 import Domain.TokenService;
 import Infrastructure.MemoryUserRepository;
